requirements.txt
beir
rank-bm25
sentence-transformers
faiss-cpu
numpy
scipy
torch
